Move SectionContainer to layout folder and update some import paths
