+++ next/src/components/common/Footer/MobileFooter.tsx
@@ -12,7 +12,7 @@ import {
   FooterLanguageSwitcher,
   FooterSocialLinks,
 } from '@/src/components/common/Footer/FooterShared'
-import SectionContainer from '@/src/components/common/SectionContainer/SectionContainer'
+import SectionContainer from '@/src/components/layouts/SectionContainer'
 import { useGeneralContext } from '@/src/components/providers/GeneralContextProvider'
 import { isDefined } from '@/src/utils/isDefined'
