+++ next/src/components/common/NavBar/AlertBanner.tsx
@@ -5,8 +5,8 @@ import { AlertIcon, CrossIcon } from 'src/assets/icons'
 import { useLocalStorage } from 'usehooks-ts'
 
 import Button from '@/src/components/common/Button/Button'
-import SectionContainer from '@/src/components/common/SectionContainer/SectionContainer'
 import Markdown from '@/src/components/formatting/Markdown/Markdown'
+import SectionContainer from '@/src/components/layouts/SectionContainer'
 import { client } from '@/src/services/graphql/gql'
 import cn from '@/src/utils/cn'
 import { useLocale } from '@/src/utils/useLocale'
