+++ next/src/components/common/NavBar/NavBarHeader/NavBarHeader.tsx
@@ -3,8 +3,8 @@ import { SearchIcon } from 'src/assets/icons'
 import Brand from '@/src/components/common/Brand/Brand'
 import Button from '@/src/components/common/Button/Button'
 import MLink from '@/src/components/common/MLink/MLink'
-import SectionContainer from '@/src/components/common/SectionContainer/SectionContainer'
 import SkipToContentButton from '@/src/components/common/SkipToContentButton/SkipToContentButton'
+import SectionContainer from '@/src/components/layouts/SectionContainer'
 import { useGeneralContext } from '@/src/components/providers/GeneralContextProvider'
 import { useLocalizations } from '@/src/components/providers/LocalizationsProvider'
 import cn from '@/src/utils/cn'
