+++ next/src/components/page-contents/DocumentPageContent.tsx
@@ -7,7 +7,7 @@ import FileRowCard from '@/src/components/cards/FileRowCard'
 import Breadcrumbs, { Breadcrumb } from '@/src/components/common/Breadcrumbs/Breadcrumbs'
 import HorizontalDivider from '@/src/components/common/Divider/HorizontalDivider'
 import DocumentPageHeader from '@/src/components/common/PageHeader/DocumentPageHeader'
-import SectionContainer from '@/src/components/common/SectionContainer/SectionContainer'
+import SectionContainer from '@/src/components/layouts/SectionContainer'
 import { useGeneralContext } from '@/src/components/providers/GeneralContextProvider'
 import { DocumentEntityFragment } from '@/src/services/graphql'
 import { formatDate } from '@/src/utils/formatDate'
