+++ next/src/components/page-contents/GeneralPageContent.tsx
@@ -3,8 +3,8 @@ import { useMemo } from 'react'
 
 import AliasInfoMessage from '@/src/components/common/AliasInfoMessage/AliasInfoMessage'
 import PageHeader from '@/src/components/common/PageHeader/PageHeader'
-import SectionContainer from '@/src/components/common/SectionContainer/SectionContainer'
 import PageHeaderSections from '@/src/components/layouts/PageHeaderSections'
+import SectionContainer from '@/src/components/layouts/SectionContainer'
 import Sections from '@/src/components/layouts/Sections'
 import RelatedArticlesSection from '@/src/components/sections/RelatedArticlesSection'
 import { PageEntityFragment } from '@/src/services/graphql'
