+++ next/src/components/page-contents/InbaArticlePageContent.tsx
@@ -5,8 +5,8 @@ import { Breadcrumb } from '@/src/components/common/Breadcrumbs/Breadcrumbs'
 import MLink from '@/src/components/common/MLink/MLink'
 import NarrowText from '@/src/components/common/NarrowText/NarrowText'
 import PageHeader from '@/src/components/common/PageHeader/PageHeader'
-import SectionContainer from '@/src/components/common/SectionContainer/SectionContainer'
 import Markdown from '@/src/components/formatting/Markdown/Markdown'
+import SectionContainer from '@/src/components/layouts/SectionContainer'
 import { useGeneralContext } from '@/src/components/providers/GeneralContextProvider'
 import ShareButtons from '@/src/components/sections/ShareButtons_Deprecated'
 import { InbaArticleEntityFragment } from '@/src/services/graphql'
