+++ next/src/components/page-contents/OfficialBoardDocumentPageContent.tsx
@@ -4,7 +4,7 @@ import React, { Fragment, ReactNode } from 'react'
 import FileRowCard from '@/src/components/cards/FileRowCard'
 import HorizontalDivider from '@/src/components/common/Divider/HorizontalDivider'
 import PageHeader from '@/src/components/common/PageHeader/PageHeader'
-import SectionContainer from '@/src/components/common/SectionContainer/SectionContainer'
+import SectionContainer from '@/src/components/layouts/SectionContainer'
 import { ParsedOfficialBoardDocumentDetail } from '@/src/services/ginis/types'
 import { formatDate } from '@/src/utils/formatDate'
 import { useLocale } from '@/src/utils/useLocale'
