+++ next/src/pages/vyhladavanie.tsx
@@ -4,8 +4,8 @@ import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
 import * as React from 'react'
 
 import PageHeader from '@/src/components/common/PageHeader/PageHeader'
-import SectionContainer from '@/src/components/common/SectionContainer/SectionContainer'
 import PageLayout from '@/src/components/layouts/PageLayout'
+import SectionContainer from '@/src/components/layouts/SectionContainer'
 import { GeneralContextProvider } from '@/src/components/providers/GeneralContextProvider'
 import { LocalizationsProvider } from '@/src/components/providers/LocalizationsProvider'
 import GlobalSearchSectionContent from '@/src/components/sections/SearchSection/GlobalSearchSectionContent'
